feat: room type

+++ internal/room/handler.go
@@ -1,6 +1,7 @@
 package room
 
 import (
+	"errors"
 	"fmt"
 	"github.com/gofiber/fiber/v2"
 	"test-chat/internal/auth"
@@ -14,11 +15,11 @@ func NewRoomRouter(router fiber.Router) {
 
 	roomRouter := router.Group("/room")
 	roomRouter.Post("/create", handler.authService.Middleware, handler.CreateRoom)
-	roomRouter.Get("/:id", handler.authService.Middleware, handler.GetRoom)
-	roomRouter.Get("/", handler.authService.Middleware, handler.GetRooms)
+	roomRouter.Get("/group", handler.authService.Middleware, handler.GetGroups)
+	roomRouter.Get("/friend-chat", handler.authService.Middleware, handler.GetFriendChats)
 
 	roomRouter.Post("/:id/add-member", handler.authService.Middleware, handler.AddMember)
-	roomRouter.Post("/:id/remove-member", handler.authService.Middleware, handler.RemoveMember)
+	roomRouter.Delete("/:id/remove-member", handler.authService.Middleware, handler.RemoveMember)
 	roomRouter.Get("/:id/list-members", handler.authService.Middleware, handler.GetMembers)
 }
 
@@ -44,7 +45,7 @@ func (h *Handler) CreateRoom(ctx *fiber.Ctx) error {
 		return err
 	}
 
-	room, err := h.roomService.CreateRoom(fmt.Sprint(ctx.Locals("userId")), *body)
+	room, err := h.roomService.CreateRoom(fmt.Sprint(ctx.Locals("userId")), *body, "group")
 	if err != nil {
 		return err
 	}
@@ -61,8 +62,17 @@ func (h *Handler) GetRoom(ctx *fiber.Ctx) error {
 	return response.SendSuccess(ctx, room)
 }
 
-func (h *Handler) GetRooms(ctx *fiber.Ctx) error {
-	rooms, err := h.roomService.GetRooms(fmt.Sprint(ctx.Locals("userId")))
+func (h *Handler) GetGroups(ctx *fiber.Ctx) error {
+	rooms, err := h.roomService.GetGroups(fmt.Sprint(ctx.Locals("userId")))
+	if err != nil {
+		return err
+	}
+
+	return response.SendSuccess(ctx, rooms)
+}
+
+func (h *Handler) GetFriendChats(ctx *fiber.Ctx) error {
+	rooms, err := h.roomService.GetFriendChats(fmt.Sprint(ctx.Locals("userId")))
 	if err != nil {
 		return err
 	}
@@ -71,6 +81,15 @@ func (h *Handler) GetRooms(ctx *fiber.Ctx) error {
 }
 
 func (h *Handler) GetMembers(ctx *fiber.Ctx) error {
+	// Check if room exists
+	roomFound, err := h.roomService.GetRoom(fmt.Sprint(ctx.Locals("userId")), ctx.Params("id"))
+	if err != nil {
+		return err
+	}
+	if roomFound == nil {
+		return response.BadRequest(ctx, errors.New("room not found"), nil)
+	}
+
 	members, err := h.roomService.GetMembers(ctx.Params("id"))
 	if err != nil {
 		return err
@@ -86,15 +105,30 @@ func (h *Handler) AddMember(ctx *fiber.Ctx) error {
 	}
 
 	// Check if user exists
-	_, err := h.userService.GetUser(body.MemberId)
+	roomFound, err := h.userService.GetUser(body.MemberId)
 	if err != nil {
 		return err
 	}
+	if roomFound == nil {
+		return response.BadRequest(ctx, errors.New("user not found"), nil)
+	}
 
 	// Check if room exists
-	_, err = h.roomService.GetRoom(fmt.Sprint(ctx.Locals("userId")), ctx.Params("id"))
+	userFound, err := h.roomService.GetRoom(fmt.Sprint(ctx.Locals("userId")), ctx.Params("id"))
 	if err != nil {
 		return err
+	}
+	if userFound == nil {
+		return response.BadRequest(ctx, errors.New("room not found"), nil)
+	}
+
+	// Check if user is already a member
+	member, err := h.roomService.GetMember(ctx.Params("id"), body.MemberId)
+	if err != nil {
+		return err
+	}
+	if member != nil {
+		return response.BadRequest(ctx, errors.New("user is already a member"), nil)
 	}
 
 	err = h.roomService.AddMember(body.MemberId, ctx.Params("id"))
@@ -111,10 +145,14 @@ func (h *Handler) RemoveMember(ctx *fiber.Ctx) error {
 		return err
 	}
 
-	// Check if room exists
-	_, err := h.roomService.GetRoom(fmt.Sprint(ctx.Locals("userId")), ctx.Params("id"))
+	// Check if member and room exists
+	member, err := h.roomService.GetMember(ctx.Params("id"), body.MemberId)
 	if err != nil {
 		return err
+	}
+
+	if member == nil {
+		return response.BadRequest(ctx, errors.New("member or room not found"), nil)
 	}
 
 	err = h.roomService.RemoveMember(body.MemberId, ctx.Params("id"))
